requirement.txt
fastapi>=0.101.0
uvicorn>=0.23.2
httpx>=0.24.1
python-dotenv>=1.0.0
pydantic>=2.3.0
langchain>=0.0.265
langgraph>=0.0.15
openai>=1.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.1